refactor: Improve reactable implementations
