--- Testing/CICDTests/Reactables/ProductNameReactableTests.cs
@@ -1,35 +1,0 @@
-﻿// <copyright file="ProductNameReactableTests.cs" company="KinsonDigital">
-// Copyright (c) KinsonDigital. All rights reserved.
-// </copyright>
-
-using CICDSystem.Reactables;
-using CICDSystem.Reactables.Core;
-using Moq;
-using Xunit;
-
-namespace CICDSystemTests.Reactables;
-
-/// <summary>
-/// Tests the <see cref="ProductNameReactable"/> class.
-/// </summary>
-public class ProductNameReactableTests
-{
-    #region Method Tests
-    [Fact]
-    public void PushNotification_WhenInvoked_SendsPushNotification()
-    {
-        // Arrange
-        var reactor = new Mock<IReactor<string>>();
-        const string expected = "test-product";
-
-        var reactable = new ProductNameReactable();
-        reactable.Subscribe(reactor.Object);
-
-        // Act
-        reactable.PushNotification(expected);
-
-        // Assert
-        reactor.Verify(m => m.OnNext(expected), Times.Once());
-    }
-    #endregion
-}
